experimental UI: move stopUpdateThread() out of synchronized, again
The experimental UI uses a thread to regularly update itself to reflect time-based
changes. As that that thread has to call synchronized methods any waiting for it to
finish has to happen outside any synchronized block. Unfortunately, 9e0308e0f7 accidentally
moved the stopUpdateThread() in buildComplete() into the synchronized block, thus giving
an opportunity for deadlocks. Move it out again.

Also, as the accounting for pending transports now happens in synchronized methods in
the state tracker, the buildEventTransportClosed() method does not have to be synchronized
any more---thus eliminating the second deadlock opportunity.

Change-Id: Icacb2ee70f4bedde1c1aac2bcfefc6fabf42fdd3
PiperOrigin-RevId: 158537844